Make chain matcher creation part of backend.
